add xander's fadetiles changes
these fadetiles can be used by scaling them to cover the region, and
better support contiguous regions of fadetiles.
i made a change where the default layer is -120, which means no create
code is required in most cases. this might break things where fadetiles
don't set their depth in create code because they assume it will be the
deefault -101, which it won't be.
otherwise, the fadetiles are designed to be backwards-compatible.