don't spawn meboids with icesync
turn meboid barriers into missile columns where it's unreasonable to
navigate around them when you have no ice beam and you're SA-X
this affects the 2 meboids in A6 that directly block progress as well as
the A5 entrance meboid and the lower A5 entrance meboid which would need
a huge detour to navigate around.
this might be expanded to all meboids in the future, depending on what
other people think.