affiliate ip with team for the life of the server
automatically sets a client's team when they connect if it's in a list
of ip's for clients who joined previously
lightly tested but it needs online testing. i have concerns about
clients syncing items with the wrong team before getting the packet that sets
them to the right team, but this cannot be tested over my same-computer
conenction.