make players reset their dmap when starting a save
had a bug where all the metroids (or maybe just all the ones killed in
the previous match) were marked as dead in the second match of the
server's uptime, and the SA-X had squares marked where neither team had
been in that match.
i was not able to replicate these in testing, but i believe this should
be the fix.

(modified from 89e0416 for cherry-picking)