fix getting frozen making you repeat actions (probably)
when frozen, the game does not update your buttons. this means that you
can get frozen with, for example, kFire being 1 and kFirePushedSteps
being 0 if you get frozen on the right frame, which causes you to
repeatedly fire. i made it so that getting frozen zeroes out the
k<Action> variables, which i beleive should solve the issue. it's not
very easy for me to test this though, so i haven't.