fix kicking players with the same IP
trying to kick players who share an IP would result in one of them
getting kicked at random. this would cause problems in tandem with
ngrok, which the server sees all ngrok players as 127.0.0.1. this fix
has it check the socket ID of th eplayer instead, and then from there
checks if they're on the kicklist.